+++ clang/docs/ReleaseNotes.rst
@@ -316,6 +316,8 @@ AST Matchers
 clang-format
 ------------
 
+- Adds ``BreakBinaryOperations`` option.
+
 libclang
 --------